move most of the image providers to a single directory
