+++ Emby.Server.Implementations/Images/CollectionFolderImageProvider.cs
