+++ Emby.Server.Implementations/Images/DynamicImageProvider.cs
