+++ Emby.Server.Implementations/Images/FolderImageProvider.cs
